refactor: adjust src code foldering
