+++ src/lib/components/MotionBox.tsx
@@ -3,7 +3,7 @@ import { chakra } from "@chakra-ui/react";
 import type { HTMLMotionProps } from "framer-motion";
 import { motion } from "framer-motion";
 
-import type { Merge } from "types/merge";
+import type { Merge } from "lib/types/merge";
 
 type MotionBoxProps = Merge<HTMLChakraProps<"div">, HTMLMotionProps<"div">>;
