+++ src/lib/pages/create/index.tsx
@@ -27,7 +27,10 @@ import { FormikErrors, useFormik } from "formik";
 import axios from "axios";
 import { useState } from "react";
 
-import { occasions, occasionsText } from "components/GreetingsTemplates/types";
+import {
+  occasions,
+  occasionsText,
+} from "lib/components/GreetingsTemplates/types";
 
 type CreateFormType = {
   name: string;
