+++ src/pages/404.ts
@@ -1,0 +1,3 @@
+import Page404 from "lib/pages/404";
+
+export default Page404;
