+++ src/pages/create.ts
@@ -1,0 +1,3 @@
+import Create from "lib/pages/create";
+
+export default Create;
