+++ src/pages/greetings/[occasion].ts
@@ -1,0 +1,3 @@
+import UnencryptedGreetingPage from "lib/pages/greetings/UnencryptedGreetingPage";
+
+export default UnencryptedGreetingPage;
