+++ src/pages/greetings/enc/[occasion].ts
@@ -1,0 +1,3 @@
+import EncryptedGreetingPage from "lib/pages/greetings/EncryptedGreetingPage";
+
+export default EncryptedGreetingPage;
