+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+import Home from "lib/pages/home";
+
+export default Home;
